+++ test/angarium/configuration_test.rb
@@ -9,7 +9,8 @@ class Angarium::ConfigurationTest < ActiveSupport::TestCase
     assert_equal :default, config.job_queue
     assert_equal 10, config.http_timeout
     assert_equal true, config.block_private_ips
-    assert_equal 5, config.retry_schedule.length
+    assert_equal 12, config.retry_schedule.length
+    assert_equal 5.seconds, config.retry_schedule.first
     assert_match(/Angarium/, config.user_agent)
   end